+++ components/Lesson/index.ts
@@ -1,0 +1,3 @@
+export { default as LearningObjectives } from "./LearningObjectives";
+export { default as PracticeQuestions } from "./PracticeQuestions";
+export { default as SummaryBox } from "./SummaryBox";