+++ components/chemistry/index.ts
@@ -1,0 +1,2 @@
+export { default as CurvedArrowGuide } from "./CurvedArrowGuide";
+export { default as ResonanceCarboxylateInteractive } from "./ResonanceCarboxylateInteractive";